// src/lib/apis/models/FinggerPlatform_Server_Shared_DataTransferObjects_Notification_NotificationInformationPublic.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { FinggerPlatform_Server_Shared_DataTransferObjects_Notification_Base_NotificationBasePublic } from './FinggerPlatform_Server_Shared_DataTransferObjects_Notification_Base_NotificationBasePublic';

export type FinggerPlatform_Server_Shared_DataTransferObjects_Notification_NotificationInformationPublic = {
  targetType?: string | null;
  hasBadge?: boolean;
  notifications?: Array<FinggerPlatform_Server_Shared_DataTransferObjects_Notification_Base_NotificationBasePublic> | null;
};